Move tool shed utility only used by tool shed into tool shed repository_util.
Doesn't belong in Galaxy utilities.
